Conflicts:
	src/entidades/EmpleadoEntity.java
	src/entityControlers/ControladorVentas.java
	src/viewControlers/ViewConsultarEmpleado.java
	src/viewControlers/ViewConsultarVentas.java
	src/viewControlers/ViewRegistrarEmpleado.java
	src/viewControlers/ViewRegistrarServicio.java
	src/vista/RegistrarServicios.fxml

+++ src/entityControlers/ControladorServicio.java
@@ -21,7 +21,7 @@ public class ControladorServicio {
         entityTransaction.commit();
     }
 
-    public static ServicioEntity crearServicio(int IDServicio, String Descripcion, Float Costo, int IDEmpleados){
+    public static ServicioEntity crearServicio(int IDServicio, String Descripcion, BigDecimal Costo, int IDEmpleados){
         ServicioEntity servicio = new ServicioEntity();
         servicio.setIdServicio(IDServicio);
         servicio.setDescripcion(Descripcion);
